Moved shape utilities to separate file
