+++ xla/service/experimental/BUILD
@@ -64,3 +64,13 @@ cc_library(
         "//xla/hlo/ir:hlo"
     ]
 )
+
+cc_library(
+    name = "shape_utils",
+    srcs = ["shape_utils.cc"],
+    hdrs = ["shape_utils.h"],
+    visibility = ["//visibility:private"],
+    deps = [
+        "//xla:shape_util"
+    ]
+)
